requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.12
Pillow==11.3.0
PyPDF2==3.0.1
google-generativeai==0.8.5
requests==2.32.5
python-dotenv==1.0.1
openai==1.51.2